Hoist config option and add a common abstract class for command block actors
