+++ worldedit-core/src/main/java/com/sk89q/worldedit/util/YAMLConfiguration.java
@@ -129,6 +129,8 @@ public class YAMLConfiguration extends LocalConfiguration {
         extendedYLimit = config.getBoolean("compat.extended-y-limit", false);
 
         setDefaultLocaleName(config.getString("default-locale", defaultLocaleName));
+
+        commandBlockSupport = config.getBoolean("command-block-support", false);
     }
 
     public void unload() {
